chore(clerk-js): Refactor eventBus to events and address PR comments
