+++ packages/clerk-js/src/core/clerk.ts
@@ -40,7 +40,6 @@ import type {
   SignOutOptions,
   SignUpProps,
   SignUpResource,
-  TokenResource,
   UnsubscribeCallback,
   UserButtonProps,
   UserProfileProps,
@@ -83,7 +82,7 @@ import {
   clerkMissingProxyUrlAndDomain,
   clerkOAuthCallbackDidNotCompleteSignInSIgnUp,
 } from './errors';
-import { eventBus } from './eventBus';
+import { eventBus, events } from './events';
 import type { FapiClient, FapiRequestCallback } from './fapiClient';
 import createFapiClient from './fapiClient';
 import {
@@ -1125,10 +1124,8 @@ export default class Clerk implements ClerkInterface {
     });
 
     // set cookie on token update
-    eventBus.addEventListener('token:update', event => {
-      const tokenEvent = event as CustomEvent;
-      const token = tokenEvent.detail.token as TokenResource;
-      this.#authService?.setAuthCookiesFromToken(token.getRawString());
+    eventBus.on(events.TokenUpdate, ({ token }) => {
+      this.#authService?.setAuthCookiesFromToken(token?.getRawString());
     });
   };
