--- packages/clerk-js/src/core/eventBus.ts
@@ -1,1 +1,0 @@
-export const eventBus = new EventTarget();
