+++ packages/clerk-js/src/core/resources/Session.ts
@@ -10,7 +10,7 @@ import type {
 import type { GetToken, GetTokenOptions, UserResource } from '@clerk/types';
 
 import { unixEpochToDate } from '../../utils/date';
-import { eventBus } from '../eventBus';
+import { eventBus, events } from '../events';
 import { SessionTokenCache } from '../tokenCache';
 import { BaseResource, Token, User } from './internal';
 
@@ -86,7 +86,7 @@ export class Session extends BaseResource implements SessionResource {
     const tokenResolver = Token.create(path);
     SessionTokenCache.set({ tokenId, tokenResolver });
     return tokenResolver.then(token => {
-      eventBus.dispatchEvent(new CustomEvent('token:update', { detail: { token } }));
+      eventBus.dispatch(events.TokenUpdate, { token });
       return token.getRawString();
     });
   };
@@ -97,7 +97,7 @@ export class Session extends BaseResource implements SessionResource {
         tokenId: this.#getCacheId(),
         tokenResolver: Promise.resolve(token),
       });
-      eventBus.dispatchEvent(new CustomEvent('token:update', { detail: { token } }));
+      eventBus.dispatch(events.TokenUpdate, { token });
     }
   };
 
@@ -134,7 +134,7 @@ export class Session extends BaseResource implements SessionResource {
       tokenResolver,
     });
     return tokenResolver.then(token => {
-      eventBus.dispatchEvent(new CustomEvent('token:update', { detail: { token } }));
+      eventBus.dispatch(events.TokenUpdate, { token });
       return token.getRawString();
     });
   };
